Code to generate and validate unsubscribe emails
This module generates and validates email addresses used in the List
Unsubscribe header. If a fixed email address is valid for unsubscribes
attackers can use it to maliciously unsubscribe others. To make this
more difficult we create a hash that is appended to the email
address. Only if this hash is valid do we accept unsubscribes.

Unsub emails are keyed by batch number, email id, and movement
name. This makes them reasonably robust to forgery, as only emails in
the same batch (about 100) use the same unsub address.